new version with new texts

+++ api/internal/v2/types/detect.go
@@ -13,7 +13,7 @@ type DetectRequest struct {
 // Optional: grade_hint (1..4), debug_reason (≤120 chars).
 type DetectResponse struct {
 	SubjectHint Subject `json:"subject_hint"`           // "math" | "russian" | "generic"
-	GradeHint   *int    `json:"grade_hint,omitempty"`   // 1..4
+	GradeHint   *int64  `json:"grade_hint,omitempty"`   // 1..4
 	Confidence  float64 `json:"confidence"`             // 0..1
 	DebugReason string  `json:"debug_reason,omitempty"` // ≤120 chars
 }
